+++ projects/ui-lib/interfaces/ui-lib-image.interface.ts
@@ -2,6 +2,7 @@ import { UiLibTextsI } from "./ui-lib-texts.interface";
 
 export interface UiLibImageI {
     url: string;
+    publicId?: string; // Used for Cloudinary images
     alt: string;
     width?: number;
     height?: number;