+++ projects/ui-lib/mapper/index.ts
@@ -1,3 +1,4 @@
 export * from './component-mappers';
 export * from './mapper.service';
 export * from './cdn.token';
+export * from './cloudinary.token';