+++ projects/ui-lib/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.28.0] - 2026-05-05
+### Added
+- add ui-image component with cloudinary integration
+
+### Changed
+- add ui-image to contact-minimal component
+
 ## [0.27.4] - 2026-05-05
 ### Changed
 - remove borders from grid guidelines